requirements.txt
mock==2.0.0; python_version < "3.3"
python-cas>=1.4.0
djangorestframework-simplejwt>=4.1.0
